Got the scoring system to work by letting the opposing the player take away points

+++ ProjectGoKart/Assets/Scripts/Scoring.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Scoring : MonoBehaviour
 {
-    public int player1score,player2score;
+    public float player1score,player2score;
 
     public GameObject checkpoint,checkpoint2,checkpoint3;
